requirements.txt
flask~=0.12.2
flask_mysqldb==0.2.0
mysqlclient~=2.0.2